+++ src/frontend/js/components/CourseProductItem/index.spec.tsx
@@ -1,9 +1,14 @@
-import { getByText, render, screen, waitFor } from '@testing-library/react';
+import {
+  getByText,
+  render,
+  screen,
+  waitFor,
+  waitForElementToBeRemoved,
+} from '@testing-library/react';
 import fetchMock from 'fetch-mock';
 import type { PropsWithChildren } from 'react';
 import { IntlProvider } from 'react-intl';
 import { QueryClientProvider } from '@tanstack/react-query';
-import faker from 'faker';
 import {
   CertificateProductFactory,
   ContextFactory as mockContextFactory,
@@ -12,9 +17,10 @@ import {
   ProductFactory,
 } from 'utils/test/factories';
 import JoanieApiProvider from 'data/JoanieApiProvider';
-import { CourseRun, Enrollment, Order, Product } from 'types/Joanie';
+import { CourseRun, Enrollment, Order, OrderState, Product } from 'types/Joanie';
 import { createTestQueryClient } from 'utils/test/createTestQueryClient';
 import { Deferred } from 'utils/test/deferred';
+import JoanieSessionProvider from 'data/SessionProvider/JoanieSessionProvider';
 import CourseProductItem from '.';
 
 jest.mock('utils/context', () => ({
@@ -58,6 +64,11 @@ describe('CourseProductItem', () => {
       style: 'currency',
     }).format(price);
 
+  beforeEach(() => {
+    fetchMock.get('https://joanie.test/api/v1.0/addresses/', []);
+    fetchMock.get('https://joanie.test/api/v1.0/credit-cards/', []);
+  });
+
   afterEach(() => {
     fetchMock.restore();
     JoanieEnrollmentFactory.afterGenerate((e: Enrollment) => e);
@@ -65,10 +76,12 @@ describe('CourseProductItem', () => {
     OrderFactory.afterGenerate((o: Order) => o);
   });
 
-  const Wrapper = ({ children }: PropsWithChildren<{}>) => (
+  const Wrapper = ({ withSession, children }: PropsWithChildren<{ withSession?: boolean }>) => (
     <IntlProvider locale="en">
       <JoanieApiProvider>
-        <QueryClientProvider client={createTestQueryClient()}>{children}</QueryClientProvider>
+        <QueryClientProvider client={createTestQueryClient({ user: withSession || null })}>
+          <JoanieSessionProvider>{children}</JoanieSessionProvider>
+        </QueryClientProvider>
       </JoanieApiProvider>
     </IntlProvider>
   );
@@ -141,31 +154,30 @@ describe('CourseProductItem', () => {
   });
 
   it('adapts information when user purchased the product', async () => {
-    const orderId = faker.datatype.uuid();
-    const product: Product = ProductFactory.afterGenerate((p: Product) => ({
-      ...p,
-      orders: [orderId],
-    })).generate();
+    const product: Product = ProductFactory.generate();
     const order: Order = OrderFactory.afterGenerate((o: Order) => ({
       ...o,
+      product: product.id,
+      course: '00000',
       target_courses: product.target_courses,
-      id: orderId,
     })).generate();
 
     fetchMock.get(`https://joanie.test/api/v1.0/products/${product.id}/?course=00000`, product);
     fetchMock.get(`https://joanie.test/api/v1.0/orders/`, [order]);
 
     render(
-      <Wrapper>
+      <Wrapper withSession>
         <CourseProductItem productId={product.id} courseCode="00000" />
       </Wrapper>,
     );
 
     // Wait for product information to be fetched
+    const loadingMessage = screen.getByRole('status', { name: 'Loading product information...' });
+    await waitForElementToBeRemoved(loadingMessage);
     await screen.findByRole('heading', { level: 3, name: product.title });
 
     // - In place of product price, a label should be displayed
-    const $enrolledInfo = screen.getByText('Enrolled');
+    const $enrolledInfo = await screen.findByText('Enrolled');
     expect($enrolledInfo.tagName).toBe('STRONG');
     expect($enrolledInfo.classList.contains('h6')).toBe(true);
 
@@ -179,7 +191,7 @@ describe('CourseProductItem', () => {
         expect($courseTitle.tagName).toBe('STRONG');
         expect($courseTitle.classList.contains('h5')).toBe(true);
         screen.getByTestId(
-          `EnrollableCourseRunList-${course.course_runs.map(({ id }) => id).join('-')}-${orderId}`,
+          `EnrollableCourseRunList-${course.course_runs.map(({ id }) => id).join('-')}-${order.id}`,
         );
       });
     });
@@ -192,11 +204,7 @@ describe('CourseProductItem', () => {
   });
 
   it('renders enrollment information when user is enrolled to a course run', async () => {
-    const orderId = faker.datatype.uuid();
-    const product: Product = CertificateProductFactory.afterGenerate((p: Product) => ({
-      ...p,
-      orders: [orderId],
-    })).generate();
+    const product: Product = CertificateProductFactory.generate();
     // - Create an order with an active enrollment
     const enrollment: Enrollment = JoanieEnrollmentFactory.afterGenerate(
       ({ state, ...e }: Enrollment): Enrollment => ({
@@ -207,8 +215,8 @@ describe('CourseProductItem', () => {
     ).generate();
     const order: Order = OrderFactory.afterGenerate((o: Order) => ({
       ...o,
-      id: orderId,
       product: product.id,
+      course: '00000',
       target_courses: product.target_courses,
       enrollments: [enrollment],
     })).generate();
@@ -217,12 +225,14 @@ describe('CourseProductItem', () => {
     fetchMock.get(`https://joanie.test/api/v1.0/orders/`, [order]);
 
     render(
-      <Wrapper>
+      <Wrapper withSession>
         <CourseProductItem productId={product.id} courseCode="00000" />
       </Wrapper>,
     );
 
     // Wait for product information to be fetched
+    const loadingMessage = screen.getByRole('status', { name: 'Loading product information...' });
+    await waitForElementToBeRemoved(loadingMessage);
     await screen.findByRole('heading', { level: 3, name: product.title });
 
     // - In place of product price, a label should be displayed
@@ -250,5 +260,54 @@ describe('CourseProductItem', () => {
         `EnrollableCourseRunList-${course.course_runs.map(({ id }) => id).join('-')}-${order.id}`,
       );
     });
+  });
+
+  it('does not render sale tunnel button if user already has a pending order', async () => {
+    const product: Product = ProductFactory.generate();
+    const order: Order = OrderFactory.afterGenerate((o: Order) => ({
+      ...o,
+      product: product.id,
+      course: '00000',
+      target_courses: product.target_courses,
+      state: OrderState.PENDING,
+    })).generate();
+
+    fetchMock.get(`https://joanie.test/api/v1.0/products/${product.id}/?course=00000`, product);
+    fetchMock.get(`https://joanie.test/api/v1.0/orders/`, [order]);
+
+    render(
+      <Wrapper withSession>
+        <CourseProductItem productId={product.id} courseCode="00000" />
+      </Wrapper>,
+    );
+
+    // Wait for product information to be fetched
+    const loadingMessage = screen.getByRole('status', { name: 'Loading product information...' });
+    await waitForElementToBeRemoved(loadingMessage);
+    await screen.findByRole('heading', { level: 3, name: product.title });
+
+    // - In place of product price, a label "Enrolled" should be displayed
+    const $enrolledInfo = await screen.findByText('Enrolled');
+    expect($enrolledInfo.tagName).toBe('STRONG');
+    expect($enrolledInfo.classList.contains('h6')).toBe(true);
+
+    // - As order is pending, the user should not be able to enroll to course runs.
+    await waitFor(() => {
+      order.target_courses.forEach((course) => {
+        const $item = screen.getByTestId(`course-item-${course.code}`);
+        // the course title shouldn't be a heading to prevent misdirection for screen reader users,
+        // but we want to it to visually look like a h5
+        const $courseTitle = getByText($item, course.title);
+        expect($courseTitle.tagName).toBe('STRONG');
+        expect($courseTitle.classList.contains('h5')).toBe(true);
+        screen.getByTestId(`CourseRunList-${course.course_runs.map(({ id }) => id).join('-')}`);
+      });
+    });
+
+    // - Render <CertificateItem />
+    screen.getByTestId('CertificateItem');
+
+    // - Does not Render <SaleTunnel />
+    expect(screen.queryByTestId('SaleTunnel')).toBeNull();
   });
 });